minor fixes

+++ inc/server/rm/rm_configs.hpp
@@ -27,7 +27,7 @@ namespace uva {
 
                         //Store the number of of weights in the reordering model entry
                         //The number of weights is a multiple of 2 
-                        static constexpr size_t NUMBER_WEIGHT_ENTRIES = 6;
+                        static constexpr uint8_t NUMBER_WEIGHT_ENTRIES = 6;
 
                         //Stores the configuration parameters for the basic reordering model implementation
                         namespace __rm_basic_model {
